+++ SmartClothingApp/src/components/DateToolbar/DateToolbar.jsx
@@ -1,0 +1,99 @@
+import React, { useEffect, useState } from "react";
+import {
+  Button,
+  View,
+  Text,
+  StyleSheet,
+  ScrollView,
+  Platform,
+} from "react-native";
+import { AppHeader } from "../../components";
+import ActivityRings from "../../components/visualizations/ActivityRings/ActivityRings";
+import { AppColor, AppStyle, AppFonts } from "../../constants/themes";
+import { useSelector, useDispatch } from "react-redux";
+import { updateActivityRings } from "../../actions/appActions";
+import DailyInsights from "../../components/DailyInsights/DailyInsights";
+import Icon from "react-native-vector-icons/FontAwesome5";
+import DateTimePicker from "@react-native-community/datetimepicker";
+import ActivityChart from "../../components/visualizations/ActivityChart/ActivityChart";
+import { useFocusEffect } from "@react-navigation/native";
+
+const DateToolbar = () => {
+  const [currentDate, setCurrentDate] = useState(new Date());
+  const [showDatePicker, setShowDatePicker] = useState(false);
+  const [selectedDay, setSelectedDay] = useState(null);
+
+  const onChangeDate = (event, selectedDate) => {
+    setShowDatePicker(Platform.OS === "ios");
+    if (selectedDate) {
+      setCurrentDate(selectedDate);
+    }
+  };
+
+  const formattedDate = currentDate.toLocaleDateString("en-US", {
+    weekday: "long",
+    month: "long",
+    day: "numeric",
+    year: "numeric",
+  });
+
+  useEffect(() => {
+    return () => {
+      setSelectedDay(null);
+    };
+  }, []);
+
+  useFocusEffect(
+    React.useCallback(() => {
+      setSelectedDay(null);
+    }, [])
+  );
+
+  return (
+    <>
+      <View style={styles.dateContainer}>
+        <Text style={styles.title}>{formattedDate}</Text>
+        <View style={styles.iconContainer}>
+          <Icon
+            name="calendar-alt"
+            size={20}
+            style={styles.icon}
+            onPress={() => setShowDatePicker(true)}
+          />
+          <Icon name="sliders-h" size={20} style={styles.icon} />
+          <Icon name="upload" size={20} style={styles.icon} />
+        </View>
+      </View>
+      {showDatePicker && (
+        <DateTimePicker
+          value={currentDate}
+          mode="date"
+          display="default"
+          onChange={onChangeDate}
+        />
+      )}
+    </>
+  );
+};
+
+const styles = StyleSheet.create({
+  title: {
+    fontSize: 20,
+    textAlign: "left",
+    color: "black",
+  },
+  dateContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  iconContainer: {
+    flexDirection: "row",
+    gap: 15,
+  },
+  icon: {
+    color: AppColor.primary,
+  },
+});
+
+export default DateToolbar;
